Lightly revised LadderPlayer and LadderPlayerTest.
Simplified the LadderPlayer.playOutLadders() method so that it doesn't
keep an extra copy of the board (it plays on runnable's board) and
cleaned up the biasing code a little. Note that we bias the search tree
by incorporating 10 * (ladderLength) winning or losing playouts. Maybe
the best constant value should be determined experimentally.

Also made LadderPlayerTest test LadderPlayer with a ladder it will win
and then with a ladder it will lose. The test passes iff LadderPlayer
suggests playing the winning ladder > 50% of the time and suggests
avoiding the losing ladder > 90% of the time.