Fixed a bug in LadderPlayer that was causing an assertion error.
In playOutLadders(), we were not updating the location where the inside
player should play at each step in the playout. When we enabled
assertions with the -ea flag, this meant we were trying to play in an
occupied point. I fixed this and did some simplifying of the
playOutLadders() method as well as the LadderPlayerTest class.