Fixed our implementation of the kgs-game_over command to stop threads only instead of clearing the board and resetting the player.
This is because kgs-game_over is not only sent when the game is completely over but also when it is only temporarily stopped. The opponent
could come back and continue playing. The fact that were we clearing the board upon receiving this command meant we would make illegal moves
after the game was picked back up.